requirements.txt
Flask
Pillow
gunicorn
google-generativeai
python-dotenv
Flask-SQLAlchemy
psycopg2
